requirements.txt
Django==3.1.7
django-allauth==0.44.0
django-anymail==8.2
django-extensions==3.1.1
django-imagekit==4.0.2
django-rest-auth==0.9.5
django-utils-six==2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.3
Pillow==8.1.2
python-dotenv==0.15.0
python-linkedin-v2==0.9.4